Minor fixes

+++ react-client/src/components/common/UserMenu/UserMenu.js
@@ -1,24 +1,48 @@
-import React from 'react'
+import React, {useEffect, useState} from 'react'
 import style from './UserMenu.module.scss'
-import {Link, useNavigate} from 'react-router-dom'
+import {Link, useLocation, useNavigate} from 'react-router-dom'
 
 const UserMenu = () => {
   const navigate = useNavigate()
+  const [username, setUsername] = useState('')
+  const [logStatus, setLogStatus] = useState(false)
+  const location = useLocation()
 
   const logout = () => {
-    localStorage.clear()
-    navigate('/login')
+    if (logStatus === true) {
+      localStorage.clear()
+      getUserName()
+      window.location.reload()
+    } else {
+      localStorage.setItem('lastPage', location.pathname)
+      navigate('/admin')
+    }
   }
 
-  const username = localStorage.getItem('username')
+  const getUserName = () => {
+    const user = localStorage.getItem('username')
+    if (!user) {
+      setUsername('User Menu')
+      setLogStatus(false)
+    } else {
+      setUsername(user, setUsername)
+      setLogStatus(true)
+    }
+  }
+
+  useEffect(() => {
+    getUserName()
+  }, [])
+
   return (
     <div className={style.drop_down}>
       <button>
-        {username} <i className={style.arrow} />
+        {username}
+        <i className={style.arrow} />
       </button>
       <div className={style.list}>
         <div className={style.item} onClick={logout}>
-          logout
+          {logStatus ? 'Logout' : 'Login'}
         </div>
         <Link to="/podcast">
           <div className={style.item}>My Pod-Space</div>
